requirements.txt
Scrapy~=2.11.1
SQLAlchemy~=2.0.29
pyodbc~=5.1.0
mysql-connector-python~=8.3.0
pandas~=2.2.1
aiohttp
python-dotenv~=1.0.1
requests~=2.31.0
beautifulsoup4
pytest
loguru
PyYAML
w3lib
itemadapter
pip
fastapi>=0.110.1
uvicorn
pandas
pydantic
python-jose[cryptography]
passlib
starlette>=0.37.2
selenium>=4.21.0